collectors/reddit/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
praw==7.8.1
prawcore==2.4.0
python-dotenv==1.1.1
requests==2.32.5
update-checker==0.18.0
urllib3==2.5.0
websocket-client==1.8.0
